requirements.txt
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PySide2==5.15.0
Pillow==8.2.0
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
numpy==1.20.2
googletrans==3.1.0a0
matplotlib==3.7.1
matplotlib-inline==0.1.2
pyttsx3==2.90